bot bingx rapide/requirements.txt
ccxt
pandas
numpy
bingx
supabase
python-dotenv